requirements.txt
Django==2.0.3
django-extensions
python-dateutil
ipython
